refactor: move shared helper functions to common pkg
This moves helper functions that are shared by multiple factory
subcommands into a common package. This will make it easier to break out
the subcommands into their own dedicated packages.
