+++ factory/pkg/commands/common/review_instructions_test.go
@@ -1,4 +1,4 @@
-package commands
+package common
 
 import (
 	"reflect"
