+++ factory/pkg/commands/watch_test.go
@@ -98,74 +98,6 @@ func TestShouldRunChoreAt(t *testing.T) {
 	}
 }
 
-func TestGetReferencedIssues(t *testing.T) {
-	tests := []struct {
-		name     string
-		headRef  string
-		title    string
-		body     string
-		expected map[int]bool
-	}{
-		{
-			name:    "Branch name contains issue number",
-			headRef: "issue_8883",
-			title:   "Some PR title",
-			body:    "Some PR body",
-			expected: map[int]bool{
-				8883: true,
-			},
-		},
-		{
-			name:    "Title and body contain issue number references",
-			headRef: "my-dev-branch",
-			title:   "Fixes #8883 and #10294",
-			body:    "Resolves issue #9271 in config-connector",
-			expected: map[int]bool{
-				8883:  true,
-				10294: true,
-				9271:  true,
-			},
-		},
-		{
-			name:     "No references",
-			headRef:  "master",
-			title:    "Clean PR without issue link",
-			body:     "Just refactoring some code",
-			expected: map[int]bool{},
-		},
-		{
-			name:    "Branch with timestamp and keyword issue link without #",
-			headRef: "ada-coder-bot:issue-11414-1783386792",
-			title:   "Fixes 11414",
-			body:    "Resolves 11414 without hash",
-			expected: map[int]bool{
-				11414: true,
-			},
-		},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			pr := &githubv39.PullRequest{
-				Head: &githubv39.PullRequestBranch{
-					Ref: &tc.headRef,
-				},
-				Title: &tc.title,
-				Body:  &tc.body,
-			}
-			got := getReferencedIssues(pr)
-			if len(got) != len(tc.expected) {
-				t.Fatalf("getReferencedIssues() returned %v; want %v", got, tc.expected)
-			}
-			for num := range tc.expected {
-				if !got[num] {
-					t.Errorf("getReferencedIssues() missed expected issue %d in %v", num, got)
-				}
-			}
-		})
-	}
-}
-
 func TestGetMissingLabelsForPR(t *testing.T) {
 	tests := []struct {
 		name      string
@@ -838,69 +770,6 @@ func TestIsSandboxTaskCompleted(t *testing.T) {
 			t.Errorf("For taskType=%s, annotatedTaskType=%s, state=%s: expected completed=%v, got %v",
 				tc.taskType, tc.annotatedTaskType, tc.state, tc.expectedCompleted, completed)
 		}
-	}
-}
-
-func TestFindWorkflowPath(t *testing.T) {
-	tests := []struct {
-		name     string
-		body     string
-		expected string
-	}{
-		{
-			name:     "clean URL",
-			body:     "Workflow: https://raw.githubusercontent.com/gke-labs/gemini-for-kubernetes-development/main/.agents/workflows/kcc-greenfield.txt",
-			expected: "https://raw.githubusercontent.com/gke-labs/gemini-for-kubernetes-development/main/.agents/workflows/kcc-greenfield.txt",
-		},
-		{
-			name:     "URL with literal escaped newline \\n",
-			body:     "This issue is to track Greenfield.\n\nWorkflow: https://raw.githubusercontent.com/gke-labs/gemini-for-kubernetes-development/main/.agents/workflows/kcc-greenfield.txt\\n",
-			expected: "https://raw.githubusercontent.com/gke-labs/gemini-for-kubernetes-development/main/.agents/workflows/kcc-greenfield.txt",
-		},
-		{
-			name:     "quoted double-quote URL should be ignored",
-			body:     "Follow workflow at \"https://raw.githubusercontent.com/gke-labs/gemini-for-kubernetes-development/main/.agents/workflows/kcc-greenfield.txt\", please.",
-			expected: "",
-		},
-		{
-			name:     "quoted single-quote URL should be ignored",
-			body:     "Check 'https://raw.githubusercontent.com/gke-labs/gemini-for-kubernetes-development/main/.agents/workflows/kcc-greenfield.txt'",
-			expected: "",
-		},
-		{
-			name:     "backticked URL should be ignored",
-			body:     "See `https://raw.githubusercontent.com/gke-labs/gemini-for-kubernetes-development/main/.agents/workflows/kcc-greenfield.txt`",
-			expected: "",
-		},
-		{
-			name:     "local relative workflow file path",
-			body:     "Please use .agents/workflows/kcc-greenfield.txt for this issue",
-			expected: ".agents/workflows/kcc-greenfield.txt",
-		},
-		{
-			name:     "local workflow path with escaped newline",
-			body:     "Workflow: .agents/workflows/kcc-greenfield.txt\\n",
-			expected: ".agents/workflows/kcc-greenfield.txt",
-		},
-		{
-			name:     "backticked local workflow path should be ignored",
-			body:     "Reference `.agents/workflows/kcc-greenfield.txt` in docs",
-			expected: "",
-		},
-		{
-			name:     "no workflow referenced",
-			body:     "Regular bug report with some code snippets",
-			expected: "",
-		},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			got := findWorkflowPath(tc.body)
-			if got != tc.expected {
-				t.Errorf("findWorkflowPath(%q) = %q; want %q", tc.body, got, tc.expected)
-			}
-		})
 	}
 }
